third_party/WebKit/LayoutTests/accessibility/aom-computed-relation-accessors.html

<!DOCTYPE HTML>
<script src="../resources/gc.js"></script>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<!--

Accessibility Object Model
Explainer: https://github.com/WICG/aom/blob/master/explainer.md
Spec: https://wicg.github.io/aom/spec/

-->

<div role="listbox" id="listbox" aria-label="container">
    <div role="option" id="option1">Option 1</div>
    <div role="option" id="option2">Option 2</div>
    <div role="option" id="option3">Option 3</div>
</div>

<script>

test(function(t) {
    assert_true(internals.runtimeFlags.accessibilityObjectModelEnabled);
}, "Make sure that Accessibility Object Model is enabled");

async_test(async function(test) {
    var element = document.getElementById("option1");
    var computedAXNode = await window.getComputedAccessibleNode(element);
    var computedParent = computedAXNode.parent;
    assert_false(computedParent === null);
    assert_equals(computedParent.name, "container");
    assert_true(computedParent.firstChild === computedAXNode);

    // Parent of listbox should be null.
    assert_equals(computedParent.parent, null);
    test.done();
}, "ComputedAccessibleNode.parent.");

async_test(async function(test) {
    var element = document.getElementById("listbox");
    var computedAXNode = await window.getComputedAccessibleNode(element);
    var computedChild = computedAXNode.firstChild
    assert_false(computedChild === null);
    assert_equals(computedChild.name, "Option 1");
    assert_true(computedChild.parent === computedAXNode);
    test.done();
}, "ComputedAccessibleNode.firstChild.");

async_test(async function(test) {
    var element = document.getElementById("listbox");
    var computedAXNode = await window.getComputedAccessibleNode(element);
    var computedChild = computedAXNode.lastChild;
    assert_false(computedChild === null);
    assert_equals(computedChild.name, "Option 3");
    assert_true(computedChild.parent === computedAXNode);
    test.done();
}, "ComputedAccessibleNode.lastChild.");

async_test(async function(test) {
    var element = document.getElementById("option2");
    var computedAXNode = await window.getComputedAccessibleNode(element);
    var computedSibling = computedAXNode.previousSibling;
    assert_false(computedSibling === null);
    assert_equals(computedSibling.name, "Option 1");

    // Check that a call to non-existant sibling is null.
    assert_equals(computedSibling.previousSibling, null);
    assert_true(computedSibling.nextSibling === computedAXNode);
    test.done();
}, "ComputedAccessibleNode.previousSibling");

async_test(async function(test) {
    var element = document.getElementById("option2");
    var computedAXNode = await window.getComputedAccessibleNode(element);
    var computedSibling = computedAXNode.nextSibling;
    assert_false(computedSibling === null);
    assert_equals(computedSibling.name, "Option 3");

    // Check that a call to non-existant sibling is null.
    assert_equals(computedSibling.nextSibling, null);
    assert_true(computedSibling.previousSibling === computedAXNode);
    test.done();
}, "ComputedAccessibleNode.nextSibling");

</script>
